requirements.txt
Pillow==9.3.0
pilmoji==2.0.2
moviepy==1.0.3


